sample/requirements.txt
pika==0.11.2
six==1.11.0
websocket-client==0.46.0
